Move ServiceMessageTypes into models
